primer avance

+++ src/context/AuthContext.js
@@ -1,5 +1,7 @@
 import React, { createContext, useEffect, useState } from 'react';
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import { auth, database, ref, set, get, child, query, orderByChild, equalTo, update, remove } from "../managers/Firebase";
+import { signInWithEmailAndPassword, signOut, reauthenticateWithCredential, EmailAuthProvider } from "firebase/auth";
 
 export const AuthContext = createContext();
 
@@ -24,18 +26,40 @@ export const AuthProvider = ({ children }) => {
     loadUser();
   }, []);
 
-  const login = async (username) => {
+  const login = async (username, ingresar) => {
+    let email = 'admin@levapanapps.com';
+    let password = '4dminLev#2406';
+
     try {
-      await AsyncStorage.setItem('username', username);
-      setUser({ username });
+      await signInWithEmailAndPassword(auth, email, password);
+      const snapshot = await get(child(ref(database), "levapan/usuarios/" + username));
+
+      console.log("snapshot", snapshot);
+
+      if (snapshot.exists()) {
+        console.log(snapshot.val())
+        await AsyncStorage.setItem('username', username);
+        setUser(snapshot.val());
+        ingresar(true);
+      }else {
+        // dispatch({ type: 'SET_USER_AUTH', payload: false });
+        // dispatch({ type: 'SET_USER', payload: [] });
+        // console.log("No hay datos disponibles");
+        // ingresar(false);
+        // return false;
+        ingresar(false);
+      }      
+      
     } catch (error) {
-      console.error('Error guardando usuario:', error);
+      console.error('Error:', error);
+      ingresar(false);
     }
   };
 
   const logout = async () => {
     try {
-      await AsyncStorage.removeItem('username');
+      await signOut(auth);
+      // await AsyncStorage.removeItem('username');
       setUser(null);
     } catch (error) {
       console.error('Error cerrando sesión:', error);
